Add Splits to RefundRequest
